chore: mv test-utils away from test cases
