+++ src/body-parser.utils.ts
@@ -1,4 +1,5 @@
 import deepMerge from 'lodash.merge';
+
 import {
   type BodyParserOptions,
   supportedBodyTypes,
